third_party/WebKit/LayoutTests/inspector/sources/debugger/internal-script-flag-for-injected-script-expected.txt

CONSOLE MESSAGE: line 1: 42
Tests that injected script has isInternalScript flag.

Internal script
Internal script
